Set uplink interface via a Hiera lookup, make Ubuntu-specific Augeas changes conditional

+++ modules/dc_profile/manifests/openstack/neutron_agent.pp
@@ -11,51 +11,37 @@
 #
 class dc_profile::openstack::neutron_agent {
 
-  # Top-level Neutron configuration common to all
   include ::neutron
   include ::neutron::plugins::ml2
   include ::neutron::agents::ml2::ovs
 
-  # A bug in the ML2 plugin deployment means that this file needs to be
-  # present
-  file { '/etc/default/neutron-server':
-    ensure  => present,
-  }
-
-  # We also need conntrack - it's a missing dependancy for the neutron
-  # packages
-  package { 'conntrack':
-    ensure => installed,
-  }
-
-  # If we're on a designated network node, configure the various
-  # additional Neutron agents for L3, DHCP and metadata functionality
-  # Note: network_node is defined at Host Group level via Foreman
   if $::network_node {
     # We want to disable GRO on the external interface
     # See: http://docs.openstack.org/havana/install-guide/install/apt/content/install-neutron.install-plug-in.ovs.html
     # Physical interface plumbed into external network
-    $uplink_if = 'p2p2'
+    $uplink_if = hiera(network_node_extif)
 
-    include ethtool
+    include ::ethtool
     ethtool { $uplink_if:
       gro => 'disabled',
     }
 
     # Ensure configuration is in place so that the external (bridged)
     # is actually brought up at boot
-    augeas { $uplink_if:
-      context => '/files/etc/network/interfaces',
-      changes => [
-          "set iface[. = '${uplink_if}'] ${uplink_if}",
-          "set iface[. = '${uplink_if}']/family inet",
-          "set iface[. = '${uplink_if}']/method manual",
-          "set iface[. = '${uplink_if}'] ${uplink_if}",
-          # Now set via DHCP
-          "rm iface[. = '${uplink_if}']/pre-up 'ip link set ${uplink_if} mtu 9000'",
-          "set iface[. = '${uplink_if}']/up 'ip link set dev ${uplink_if} up'",
-          "set iface[. = '${uplink_if}']/down 'ip link set dev ${uplink_if} down'",
-      ],
+    if $::osfamily == 'Debian' {
+      augeas { $uplink_if:
+        context => '/files/etc/network/interfaces',
+        changes => [
+            "set iface[. = '${uplink_if}'] ${uplink_if}",
+            "set iface[. = '${uplink_if}']/family inet",
+            "set iface[. = '${uplink_if}']/method manual",
+            "set iface[. = '${uplink_if}'] ${uplink_if}",
+            # Now set via DHCP
+            "rm iface[. = '${uplink_if}']/pre-up 'ip link set ${uplink_if} mtu 9000'",
+            "set iface[. = '${uplink_if}']/up 'ip link set dev ${uplink_if} up'",
+            "set iface[. = '${uplink_if}']/down 'ip link set dev ${uplink_if} down'",
+        ],
+      }
     }
 
     # Set default domain
@@ -72,7 +58,7 @@ class dc_profile::openstack::neutron_agent {
 
     unless $::is_vagrant {
       if $environment == 'production' {
-        include dc_logstash::client::neutron
+        include ::dc_logstash::client::neutron
       }
     }
